Break BasePage up into several mixin classes and create a new NonInternationalBasePage class to use as a base for non-international page types
